+++ CHANGELOG.md
@@ -5,14 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0-a3] - 2025-10-03
+## [0.4.1-a1] - 2025-10-06
 
 ### Added
 - Automation for preparing a new release version: build_release.py
 - Checklist for building a release: docs/development/release_checklist.md
+- VBA trust access check at start of test execution. Can be skipped with option --skip-vba-check
+- Test module for edit command
+- First section of a user manual, listing command line options
 
 ### Changed
 - Refactoring: Common arguments are now handled centrally in office_cli.py -> excel_vba.py and access_vba.py only contain specialties
+- Roadmap updated
 
 ### Fixed
 - proper inclusion of version number in standalone executables